+++ utils/process_ticker_10k_data.py
@@ -1,0 +1,73 @@
+from utils.get_ticker_10k_filings import get_ticker_10k_filings
+from utils.collect_ticker_files import collect_ticker_files
+from utils.delete_txt_files import delete_txt_files
+from utils.parse_html_file_mda import parse_html_file_mda
+from shutil import rmtree
+from dotenv import load_dotenv
+import os
+from supabase import create_client
+
+# Supabase API keys
+load_dotenv()
+SUPABASE_URL = os.environ["SUPABASE_URL"]
+SUPABASE_KEY = os.environ["SUPABASE_KEY"]
+Client = create_client(SUPABASE_URL, SUPABASE_KEY)
+
+
+def process_ticker_10k_data(ticker):
+    try:
+        get_ticker_10k_filings(ticker)
+    except Exception as e:
+        print(f"Error occurred while downloading filings for {ticker}: {e}")
+        return {}
+
+    ticker_files_dict = collect_ticker_files()
+    delete_txt_files(ticker_files_dict.get(ticker, []))
+
+    all_parsed_data = {}
+    for html_file in ticker_files_dict.get(ticker, []):
+        if html_file.endswith(".html"):
+            path_parts = html_file.split("/")
+            cik_year_acc = path_parts[4].split("-")
+
+            if len(cik_year_acc) < 3:
+                print(f"Skipping file with unexpected format: {html_file}")
+                continue
+
+            CIK = cik_year_acc[0]
+            # Convert the two-digit year to four digits
+            two_digit_year = cik_year_acc[1]
+            if (
+                int(two_digit_year) > 50
+            ):  # Assuming we're starting from 1950 for simplicity
+                Year = "19" + two_digit_year
+            else:
+                Year = "20" + two_digit_year
+            AccessionNumber = cik_year_acc[2]
+            try:
+                parsed_data = parse_html_file_mda(html_file)
+            except Exception as e:
+                print(f"Could not parse {html_file} due to error: {e}")
+                continue
+
+            try:
+                filing_dict = {
+                    "ticker": ticker,
+                    "cik": CIK,
+                    "year": int(Year),
+                    "accession_number": AccessionNumber,
+                    "mda_section": parsed_data.get("MD&A", "Section not found"),
+                    "target_word_frequency": parsed_data.get("word_frequency", "{}"),
+                }
+            except ValueError:
+                print(f"Skipping file with invalid year format in {html_file}")
+                continue
+
+            all_parsed_data[AccessionNumber] = filing_dict
+
+    all_parsed_data_list = list(all_parsed_data.values())
+    new_10k_reports_to_supabase_mda(all_parsed_data_list, Client)
+    # Clear the data folder after processing
+    if os.path.exists("data"):
+        rmtree("data")
+    return all_parsed_data
